updated aircraft to show empty model, deleted seat class, some code suggestions
