+++ src/main/java/com/keyin/rest/aircraft/Aircraft.json
@@ -2,141 +2,181 @@
   {
     "aircraft_ID": 1,
     "model": "Boeing 737",
-    "manufacturer": "Boeing",
     "airline": "Air Canada",
-    "capacity": 160
+    "capacity": 160,
+    "rows": 27,
+    "columns": 6,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 2,
     "model": "Airbus A320",
-    "manufacturer": "Airbus",
     "airline": "WestJet",
-    "capacity": 180
+    "capacity": 180,
+    "rows": 30,
+    "columns": 6,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 3,
     "model": "Bombardier Dash 8 Q400",
-    "manufacturer": "De Havilland Canada (Bombardier)",
     "airline": "Air Canada Express",
-    "capacity": 78
+    "capacity": 78,
+    "rows": 20,
+    "columns": 4,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 4,
     "model": "Boeing 777",
-    "manufacturer": "Boeing",
     "airline": "Air Canada",
-    "capacity": 300
+    "capacity": 300,
+    "rows": 50,
+    "columns": 10,
+    "numAisles": 2
   },
   {
     "aircraft_ID": 5,
     "model": "Boeing 787 Dreamliner",
-    "manufacturer": "Boeing",
     "airline": "Air Canada",
-    "capacity": 250
+    "capacity": 250,
+    "rows": 43,
+    "columns": 9,
+    "numAisles": 2
   },
   {
     "aircraft_ID": 6,
     "model": "Airbus A220",
-    "manufacturer": "Airbus (formerly Bombardier CSeries)",
     "airline": "Air Canada",
-    "capacity": 135
+    "capacity": 135,
+    "rows": 25,
+    "columns": 5,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 7,
     "model": "Boeing 767",
-    "manufacturer": "Boeing",
     "airline": "WestJet",
-    "capacity": 250
+    "capacity": 250,
+    "rows": 42,
+    "columns": 8,
+    "numAisles": 2
   },
   {
     "aircraft_ID": 8,
     "model": "Airbus A330",
-    "manufacturer": "Airbus",
     "airline": "Air Transat",
-    "capacity": 335
+    "capacity": 335,
+    "rows": 60,
+    "columns": 9,
+    "numAisles": 2
   },
   {
     "aircraft_ID": 9,
     "model": "Embraer E175",
-    "manufacturer": "Embraer",
     "airline": "Air Canada Express",
-    "capacity": 76
+    "capacity": 76,
+    "rows": 19,
+    "columns": 4,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 10,
     "model": "Bombardier CRJ Series",
-    "manufacturer": "Bombardier",
     "airline": "Air Canada Jazz",
-    "capacity": 50
+    "capacity": 50,
+    "rows": 13,
+    "columns": 4,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 11,
     "model": "Boeing 737 MAX",
-    "manufacturer": "Boeing",
     "airline": "WestJet",
-    "capacity": 172
+    "capacity": 172,
+    "rows": 29,
+    "columns": 6,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 12,
     "model": "Airbus A321",
-    "manufacturer": "Airbus",
     "airline": "Air Canada Rouge",
-    "capacity": 200
+    "capacity": 200,
+    "rows": 34,
+    "columns": 6,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 13,
     "model": "Boeing 747",
-    "manufacturer": "Boeing",
     "airline": "Air Canada Cargo",
-    "capacity": 400
+    "capacity": 400,
+    "rows": 66,
+    "columns": 10,
+    "numAisles": 2
   },
   {
     "aircraft_ID": 14,
     "model": "Airbus A319",
-    "manufacturer": "Airbus",
     "airline": "Air Canada",
-    "capacity": 144
+    "capacity": 144,
+    "rows": 24,
+    "columns": 6,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 15,
     "model": "Boeing 757",
-    "manufacturer": "Boeing",
     "airline": "Air Canada Rouge",
-    "capacity": 200
+    "capacity": 200,
+    "rows": 38,
+    "columns": 6,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 16,
     "model": "Bombardier Global 6000",
-    "manufacturer": "Bombardier",
     "airline": "Air Canada Jetz",
-    "capacity": 13
+    "capacity": 13,
+    "rows": 4,
+    "columns": 2,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 17,
     "model": "Airbus A330neo",
-    "manufacturer": "Airbus",
     "airline": "Air Transat",
-    "capacity": 375
+    "capacity": 375,
+    "rows": 63,
+    "columns": 9,
+    "numAisles": 2
   },
   {
     "aircraft_ID": 18,
     "model": "Boeing 767 Freighter",
-    "manufacturer": "Boeing",
     "airline": "Cargojet Airways",
-    "capacity": 420
+    "capacity": 420,
+    "rows": 70,
+    "columns": 9,
+    "numAisles": 2
   },
   {
     "aircraft_ID": 19,
     "model": "De Havilland Canada Dash 7",
-    "manufacturer": "De Havilland Canada",
     "airline": "Air BC",
-    "capacity": 50
+    "capacity": 50,
+    "rows": 12,
+    "columns": 4,
+    "numAisles": 1
   },
   {
     "aircraft_ID": 20,
     "model": "Cessna Citation X",
-    "manufacturer": "Cessna",
     "airline": "Private/Corporate",
-    "capacity": 8
+    "capacity": 8,
+    "rows": 2,
+    "columns": 2,
+    "numAisles": 1
   }
 ]
